Get entity children

Get all children for a given entity document

Usage

get_entity_children(entity_id)

Arguments

entity_id

Either the GUID or the slug string for a given entity

Value

A data frame with details about the given entity's children

Details

This is recursive, so a destination will return all parks and all rides within those parks.

Examples

## Entity children data for Walt Disney World
# \donttest{
get_entity_children("waltdisneyworldresort")
#> # A tibble: 518 × 12
#>    id                       name  entity_type timezone children_id children_name
#>    <chr>                    <chr> <chr>       <chr>    <chr>       <chr>        
#>  1 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 7f6e15e9-f… Hoop-Dee-Doo…
#>  2 e957da41-3552-4cf6-b636… Walt… DESTINATION America… d7618a38-f… Amare        
#>  3 e957da41-3552-4cf6-b636… Walt… DESTINATION America… a08d9f84-f… DiVine       
#>  4 e957da41-3552-4cf6-b636… Walt… DESTINATION America… d90f0556-5… Stir         
#>  5 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 51392ca4-f… Main Street …
#>  6 e957da41-3552-4cf6-b636… Walt… DESTINATION America… d3461b18-2… Pizza al Tag…
#>  7 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 796b0a25-c… Jungle Cruise
#>  8 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 99fc0f7d-6… Blue Ribbon …
#>  9 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 15107c34-6… Space 220 Lo…
#> 10 e957da41-3552-4cf6-b636… Walt… DESTINATION America… 7ad20d41-f… Planet Holly…
#> # ℹ 508 more rows
#> # ℹ 6 more variables: children_entity_type <chr>, children_slug <chr>,
#> #   children_parent_id <chr>, children_external_id <chr>,
#> #   children_location_longitude <dbl>, children_location_latitude <dbl>
# }