Get entity live data

Get an entity's live data (wait times, availability, parade times, etc.) as well as for all child entities.

Usage

get_entity_live_data(entity_id, verbose = TRUE)

Arguments

entity_id

Character. Either the GUID or the slug string for a given entity.

verbose

Logical. Whether you would like messages to print in the console.

Value

A data frame with live data for the entity

Examples

## Live data for Walt Disney World
# \donttest{
get_entity_live_data("waltdisneyworldresort")
#> ℹ Pulling live data for Walt Disney World® Resort
#> # A tibble: 272 × 27
#>    id                  name  entity_type park_id external_id status last_updated
#>    <chr>               <chr> <chr>       <chr>   <chr>       <chr>  <chr>       
#>  1 bc1ffa86-9b1a-4ce9… Tree… ATTRACTION  1c84a2… 80010217;e… OPERA… 2022-09-30T…
#>  2 ead53ea5-22e5-4095… Disn… PARK        NA      80007834;e… CLOSED 2024-06-07T…
#>  3 ada6a611-31ef-426a… Runo… ATTRACTION  ead53e… 80010183;e… OPERA… 2025-02-25T…
#>  4 2a55aee6-c87a-4bda… Tree… SHOW        1c84a2… 18447293;e… OPERA… 2025-02-24T…
#>  5 55c531b6-3ce2-4c47… DINO… ATTRACTION  1c84a2… 80010123;e… OPERA… 2025-02-25T…
#>  6 96455de6-f4f1-403c… The … ATTRACTION  75ea57… 80010210;e… OPERA… 2025-02-25T…
#>  7 888fb4a4-7adf-47a1… Main… ATTRACTION  75ea57… 80010165;e… CLOSED 2024-07-26T…
#>  8 0d94ad60-72f0-4551… The … ATTRACTION  75ea57… 80010213;e… OPERA… 2025-02-25T…
#>  9 b61e138d-d0bd-4b3a… Narc… RESTAURANT  NA      90001983;e… OPERA… 2024-07-26T…
#> 10 f8d2730d-1fcb-414e… Dona… SHOW        1c84a2… 18984704;e… CLOSED 2024-06-09T…
#> # ℹ 262 more rows
#> # ℹ 20 more variables: operating_hours <list>, showtimes <list>,
#> #   forecast <list>, dining_availability <list>, queue_standby_wait_time <int>,
#> #   queue_return_time_state <chr>, queue_return_time_return_end <lgl>,
#> #   queue_return_time_return_start <chr>, queue_paid_return_time_state <chr>,
#> #   queue_paid_return_time_return_end <lgl>,
#> #   queue_paid_return_time_return_start <lgl>, …
# }